requirements.txt
flask==2.1.1
flask_cors==3.0.10
mysql-connector-python==8.0.23
python-dotenv==0.19.2
Werkzeug==2.2.2  # Pin Werkzeug to a compatible version
